userservice/src/test/java/io/wiiisdom/userservice/controllers/UserControllerTest.java
package io.wiiisdom.userservice.controllers;

import io.wiiisdom.userservice.entities.User;
import io.wiiisdom.userservice.services.UserService;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;
import org.modelmapper.ModelMapper;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.context.TestConfiguration;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.context.annotation.Bean;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;

import java.util.Arrays;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;


@WebMvcTest(UserController.class)
public class UserControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserService userService;

    @TestConfiguration
    static class TestConfig {
        @Bean
        public ModelMapper modelMapper() {
            return new ModelMapper();
        }
    }

    @Test
    public void getUsersShouldReturnUsersList() throws Exception {
        Iterable<User> users = Arrays.asList(
                new User("azerty@gmail.com", "mickey"),
                new User("dour@gmail.com", "globule")
        );
        when(userService.getUsers()).thenReturn(users);

        mockMvc.perform(MockMvcRequestBuilders.get("/users"))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.users[0].email").value("azerty@gmail.com"))
                .andExpect(jsonPath("$.users[1].email").value("dour@gmail.com"));
    }
}
